feat: Complete StageSelectScene.js splitting (Task 8)
- Refactored StageSelectScene.js to lightweight Main Controller (2,573→331 words)
- Created EventStageDataManager.js with event functionality (1,655 words)
- Created StageSelectDataManager.js with stage management (1,127 words)
- Applied Main Controller Pattern with dependency injection
- Maintained backward compatibility and public API
- All files now under 2,500 words for MCP tool compatibility
